swp-04 - Adjusting App structure for the Browser route Component
