+++ src/App.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+
+import Routes from './Routes';
+
+function App() {
+  return(
+    <BrowserRouter>
+      <Routes />
+    </BrowserRouter>
+  );
+}
+
+export default App;
