+++ src/components/AlbumSection/RecentSearch/AlbumItem.js
@@ -1,12 +1,15 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useRouteMatch } from 'react-router-dom';
 
 const AlbumItem = (props) => {
+    let match = useRouteMatch();
+
     const { name, artist, images, id } = props.album;
     const cover = images.length ? images[1].url : '';
 
     return(
-        <Link to={`/${id}`} >
+        // <Link to={`${match.url}/${id}`} >
+        <Link to={`/#${id}`} >
             <div>
                 <img className="album-cover" src={ cover } />
                 <p className="album-name">{ name }</p>
